feat(chess): implement chess board and piece management classes
- Introduced `ChessBoard` class to manage the chess game state, including piece positions and actions such as move, capture, and promote.
- Added `PiecePositions` class to track live positions of pieces for each player, enhancing move and capture efficiency.
- Implemented methods for retrieving piece information based on coordinates and handling game actions in a structured manner.
- Enhanced the overall architecture for better performance and maintainability in chess game processing.
